Problem: some compilers don't like `sqlite::flat_size`
```
In member function ‘std::size_t sqlite::flat_size()’:
extensions/omni_sqlite/sqlite.cpp:9:58: error: invalid conversion from ‘int64_t*’ {aka ‘long int*’} to ‘sqlite3_int64*’ {aka ‘long long int*’} [-fpermissive]
    9 |   unsigned char *v = sqlite3_serialize(db.get(), "main", &_flat_size, 0);
```

Solution: re-type it

+++ extensions/omni_sqlite/sqlite.cpp
@@ -12,7 +12,7 @@ std::size_t sqlite::flat_size() {
         std::format("Failed to serialize SQLite: {}", sqlite3_errmsg(db.get())));
   }
 
-  return _flat_size;
+  return static_cast<std::size_t>(_flat_size);
 }
 
 void sqlite::flatten_into(std::span<std::byte> buffer) {
